added comment and like functionality

+++ src/pages/DeleteBlog.js
@@ -7,18 +7,20 @@ import { Button } from "react-bootstrap";
 
 const DeleteBlog = ({ id, imageUrl }) => {
   const handleDelete = async () => {
-    try {
-      await deleteDoc(doc(db, "BlogsData", id));
-      toast.info("Successfully deleted blog!");
-      const storageRef = ref(storage, imageUrl);
-      await deleteObject(storageRef);
-    } catch (error) {
-      toast.error("Fail to delete blog, please try again.");
+    if (window.confirm("Wanna delete ?")) {
+      try {
+        await deleteDoc(doc(db, "BlogsData", id));
+        toast.info("Successfully deleted blog!");
+        const storageRef = ref(storage, imageUrl);
+        await deleteObject(storageRef);
+      } catch (error) {
+        toast.error("Fail to delete blog, please try again.");
+      }
     }
   };
 
   return (
-    <Button variant="danger" onClick={handleDelete}>
+    <Button className="mt-5" variant="danger" onClick={handleDelete}>
       Delete
     </Button>
   );
